chore: make cloud-pr use the branch name (gen1) (#3350) (#3357)
`yarn cloud-pr` currently asks you for the PR number, which is annoying
if you need to run it a lot because you need to input the PR number over
and over again.

Instead, it can work exactly the same as `yarn cloud-e2e`: using the
current branch name instead. That way you can submit it unattended.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.