feat: ElasticSearch server-side encryption (#3339)
"node to node" encryption is currently configurable for ElasticSearch
domains, but not "encryption at rest".

Add a parameter for that option, and turn it on by default for E2E
tests.

----

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Momo Kornher <mail@moritzkornher.de>