chore: add a timeout on the entire CLI invocation (#3389)
In ca-central-1 our canary is taking more than an hour to finih. It's
not producing no output for 5 minutes because the "no output timeout"
would have caught that.

Add another "total duration timeout", set to 30 minutes, to see what
it's doing as it's timing out.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.